cleanup System/Sync
- ASSERT_SYNCED instead of ASSERT_SYNCED_PRIMITIVE & ASSERT_SYNCED_FLOAT3
- removed never used 64 bit int support from Upcast and SyncedPrimitive
- removed all binary operator overloads from SyncedPrimitive   (implicit
    conversions seem to work fine without, added test for this)
- removed custom std::min and std::max  (may result in undefined behavior
    according to the standard, e.g. 17.6.4.2.1 in C++0x draft)
- less friends, more public functions
- missing break in HsiehHash
- constify some pointers

+++ rts/Game/NetCommands.cpp
@@ -629,7 +629,7 @@ void CGame::ClientReadNet()
 				float energyShare = std::min(*(float*)&inbuf[8], (float)team1->energy);
 
 				if (metalShare != 0.0f) {
-					metalShare = std::min(metalShare, team1->metal);
+					metalShare = std::min(metalShare, (float)team1->metal);
 					if (!luaRules || luaRules->AllowResourceTransfer(teamID1, teamID2, "m", metalShare)) {
 						team1->metal                       -= metalShare;
 						team1->metalSent                   += metalShare;
@@ -640,7 +640,7 @@ void CGame::ClientReadNet()
 					}
 				}
 				if (energyShare != 0.0f) {
-					energyShare = std::min(energyShare, team1->energy);
+					energyShare = std::min(energyShare, (float)team1->energy);
 					if (!luaRules || luaRules->AllowResourceTransfer(teamID1, teamID2, "e", energyShare)) {
 						team1->energy                       -= energyShare;
 						team1->energySent                   += energyShare;
